* cleaned up project generation and moved imgui into seperate lib for more convinient ide experience
* additionally added cmake option for out of source build

+++ source/Slipper_Engine/test/TestGui.cpp
@@ -8,8 +8,8 @@
 #include "GraphicsEngine.h"
 #include "Presentation/OffscreenSwapChain.h"
 #include "Presentation/SwapChain.h"
-#include "Window.h"
 #include "Texture/Texture2D.h"
+#include "Window.h"
 
 using namespace std::placeholders;
 
@@ -23,7 +23,8 @@ void TestGui::OnGuiRender()
 
         m_viewportImages.resize(m_graphicsEngine->viewportSwapChain->numImages);
         for (uint32_t i{0}; i < m_graphicsEngine->viewportSwapChain->numImages; i++) {
-            m_viewportImages[i] = m_graphicsEngine->viewportPresentationTextures[i]->imageInfo.view;
+            m_viewportImages[i] =
+                m_graphicsEngine->viewportPresentationTextures[i]->imageInfo.view;
         }
 
         for (const auto viewport_image : m_viewportImages) {
@@ -41,12 +42,11 @@ void TestGui::OnGuiRender()
     auto window_size = ImGui::GetContentRegionAvail();
     window_size.x = std::max(1.0f, window_size.x);
     window_size.y = std::max(1.0f, window_size.y);
-    if (last_window_size.x != window_size.x || last_window_size.y != window_size.y)
-    {
-        Slipper::Application::Get().OnViewportResize(window_size.x, window_size.y);
+    if (last_window_size.x != window_size.x || last_window_size.y != window_size.y) {
+        Slipper::Application::Get().OnViewportResize(static_cast<uint32_t>(window_size.x),
+                                                     static_cast<uint32_t>(window_size.y));
     }
-    ImGui::Image(
-        m_imguiViewportImages[current_frame], window_size);
+    ImGui::Image(m_imguiViewportImages[current_frame], window_size);
     last_window_size = window_size;
     ImGui::End();
 
@@ -57,14 +57,14 @@ void TestGui::Init()
 {
     AppComponent::Init();
 
-    Slipper::Application::Get().AddViewportResizeCallback(std::bind(&TestGui::OnViewportResize, this, _1, _2));
+    Slipper::Application::Get().AddViewportResizeCallback(
+        std::bind(&TestGui::OnViewportResize, this, _1, _2));
     m_graphicsEngine = &Slipper::GraphicsEngine::Get();
 }
 
 void TestGui::OnViewportResize(uint32_t Width, uint32_t Height)
 {
-    for (auto imgui_viewport_image : m_imguiViewportImages)
-    {
+    for (auto imgui_viewport_image : m_imguiViewportImages) {
         ImGui_ImplVulkan_RemoveTexture(imgui_viewport_image);
     }
     m_imguiViewportImages.clear();
